+++ shuup_tests/campaigns/test_basket_campaigns.py
@@ -6,29 +6,32 @@
 # LICENSE file in the root directory of this source tree.
 from decimal import Decimal
 
-import pytest
 from django.db import IntegrityError
 
+import pytest
 from shuup.campaigns.admin_module.forms import BasketCampaignForm
 from shuup.campaigns.models.basket_conditions import (
     BasketTotalAmountCondition, BasketTotalProductAmountCondition,
-    CategoryProductsBasketCondition
+    CategoryProductsBasketCondition, ProductsInBasketCondition
 )
 from shuup.campaigns.models.basket_line_effects import DiscountFromCategoryProducts
 from shuup.campaigns.models.basket_effects import (
     BasketDiscountAmount, BasketDiscountPercentage
 )
+from shuup.campaigns.models.basket_line_effects import \
+    DiscountFromCategoryProducts
 from shuup.campaigns.models.campaigns import (
     BasketCampaign, Coupon, CouponUsage
 )
-from shuup.core.models import OrderLineType, ShopProduct
+from shuup.core.models import (
+    Category, OrderLineType, Shop, ShopProduct, ShopStatus
+)
 from shuup.core.order_creator import OrderCreator
 from shuup.front.basket import get_basket
 from shuup.front.basket.commands import handle_add_campaign_code
 from shuup.testing.factories import (
-    create_product, get_default_product, get_default_supplier,
-    get_shipping_method, CategoryFactory
-)
+    get_default_supplier, get_shipping_method,
+    create_product, CategoryFactory, get_default_product, get_default_shop)
 from shuup_tests.campaigns import initialize_test
 from shuup_tests.core.test_order_creator import seed_source
 from shuup_tests.utils import printable_gibberish
@@ -364,3 +367,58 @@ def test_coupon_uniqueness(rf):
     with pytest.raises(IntegrityError):
         second_campaign.coupon = coupon
         second_campaign.save()
+
+
+@pytest.mark.django_db
+def test_product_basket_campaigns():
+    shop = get_default_shop()
+    product = create_product("test", shop, default_price=20)
+    shop_product = product.get_shop_instance(shop)
+    cat = Category.objects.create(name="test")
+    campaign = BasketCampaign.objects.create(active=True, shop=shop, name="test")
+
+    # no rules
+    assert BasketCampaign.get_for_product(shop_product).count() == 0
+
+    # category condition that doesn't match
+    cat_condition = CategoryProductsBasketCondition.objects.create(category=cat)
+    campaign.conditions.add(cat_condition)
+    assert BasketCampaign.get_for_product(shop_product).count() == 0
+
+    # category condition that matches
+    shop_product.categories.add(cat)
+    assert BasketCampaign.get_for_product(shop_product).count() == 1
+
+    # category effect that doesn't match
+    effect = DiscountFromCategoryProducts.objects.create(campaign=campaign, category=cat)
+    shop_product.categories.remove(cat)
+    campaign.line_effects.add(effect)
+    assert BasketCampaign.get_for_product(shop_product).count() == 0
+
+    # category effect and condition that matches
+    shop_product.primary_category = cat
+    shop_product.save()
+    assert BasketCampaign.get_for_product(shop_product).count() == 1
+
+
+@pytest.mark.django_db
+def test_product_basket_campaigns2():
+    shop = get_default_shop()
+    product = create_product("test", shop, default_price=20)
+    shop_product = product.get_shop_instance(shop)
+    campaign = BasketCampaign.objects.create(active=True, shop=shop, name="test")
+
+    condition = ProductsInBasketCondition.objects.create(quantity=1)
+    campaign.conditions.add(condition)
+    assert BasketCampaign.get_for_product(shop_product).count() == 0
+
+    condition.products.add(product)
+    assert BasketCampaign.get_for_product(shop_product).count() == 1
+
+    shop1 = Shop.objects.create(name="testshop", identifier="testshop", status=ShopStatus.ENABLED, public_name="testshop")
+    sp = ShopProduct.objects.create(product=product, shop=shop1, default_price=shop1.create_price(200))
+
+    campaign.shop = shop1
+    campaign.save()
+    assert BasketCampaign.get_for_product(shop_product).count() == 0
+    assert BasketCampaign.get_for_product(sp).count() == 1